Updated files

+++ src/main/java/com/example/demo/controllers/TaskController.java
@@ -1,5 +1,7 @@
 package com.example.demo.controllers;
 
+import com.example.demo.exception.EmptyTaskListException;
+import com.example.demo.exception.MessageNotReadableException;
 import com.example.demo.exception.ResourceNotFoundException;
 import com.example.demo.models.Task;
 import com.example.demo.services.TaskService;
@@ -10,9 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 @Controller
 @RequestMapping("/api/v1/tasks")
@@ -22,56 +22,69 @@ public class TaskController {
     private TaskService taskService;
 
     // Get all tasks - returns response entity with HTTP status code of 200 (OK) and a list of tasks
-    @GetMapping("/")
-    public ResponseEntity<List<Task>> getAllTasks() {
-        return ResponseEntity.ok(taskService.getAllTask());
-    }
+
+    // Original - GFG
 //    @GetMapping("/")
 //    public ResponseEntity<List<Task>> getAllTasks() {
-//        List <Task> taskList = taskService.getAllTask();
-//        if(!taskList.isEmpty()) {
-//            return ResponseEntity.ok(taskList);
-//        }
-//        else {
-//            return ResponseEntity.status(HttpStatus.NO_CONTENT).body(List.<Task>of());
-//        }
+//        return ResponseEntity.ok(taskService.getAllTask());
 //    }
-
+    // My version - Msg to be returned if list is empty.
+    @GetMapping("/")
+    public ResponseEntity<List<Task>> getAllTasks() throws EmptyTaskListException {
+        List <Task> taskList = taskService.getAllTask();
+        if(taskList.isEmpty()) {
+            throw new EmptyTaskListException("No task available.");
+        }
+        return ResponseEntity.ok(taskList);
+    }
 
     // Get all completed tasks - returns response entity with HTTP status code of 200 (OK) and a list of completed tasks
     @GetMapping("/completed")
-    public ResponseEntity<List<Task>> getAllCompletedTasks() {
-        return ResponseEntity.ok(taskService.findAllCompletedTask());
+    public ResponseEntity<List<Task>> getAllCompletedTasks() throws EmptyTaskListException {
+        List <Task> taskList = taskService.findAllCompletedTask();
+        if(taskList.isEmpty()) {
+            throw new EmptyTaskListException("No task available.");
+        }
+        return ResponseEntity.ok(taskList);
     }
     // Get all incomplete tasks - returns response entity with HTTP status code of 200 (OK) and a list of incomplete tasks
     @GetMapping("/incomplete")
-    public ResponseEntity<List<Task>> getAllIncompleteTasks() {
-        return ResponseEntity.ok(taskService.findAllInCompleteTask());
+    public ResponseEntity<List<Task>> getAllIncompleteTasks() throws EmptyTaskListException {
+        List <Task> taskList = taskService.findAllInCompleteTask();
+        if(taskList.isEmpty()) {
+            throw new EmptyTaskListException("No task available.");
+        }
+        return ResponseEntity.ok(taskList);
     }
     // Create task - returns response entity with HTTP status code of 200 (OK)
     @PostMapping("/")
-    public ResponseEntity<Task> createTask(@RequestBody Task task) {
-        return ResponseEntity.ok(taskService.createNewTask(task));
+    public ResponseEntity<Task> createTask(@Valid @RequestBody Task task) {
+            return new ResponseEntity<>(taskService.createNewTask(task), HttpStatus.CREATED);
     }
     // Update task - returns response entity with HTTP status code of 200 (OK)
     @PutMapping("/{id}")
-    public ResponseEntity<Task> updateTask(@PathVariable Long id,@Valid @RequestBody Task task) {
-        task.setId(id);
-        Optional<Task> updatedTask = Optional.ofNullable(taskService.updateTask(task));
-        if(!updatedTask.isPresent()) {
-            return ResponseEntity.notFound().build();
+    public ResponseEntity<Task> updateTask(@Valid @PathVariable Long id, @Valid @RequestBody Task task) {
+        // If task cannot be found by id, throw ResourceNotFoundException
+        if(taskService.findTaskById(id)==null) {
+            throw new ResourceNotFoundException(id);
         }
         else {
-            return ResponseEntity.ok(updatedTask.get());
+            task.setId(id);
+            return ResponseEntity.ok(taskService.updateTask(task));
         }
-
     }
+
     // Delete task by id - returns response entity with HTTP status code of 200 (OK) and a response body containing boolean value (true)
     @DeleteMapping("/{id}")
-    public ResponseEntity<Boolean> deleteTask(@PathVariable Long id) {
+    public ResponseEntity<Boolean> deleteTask(@Valid @PathVariable Long id) {
+        // If task cannot be found by id, throw ResourceNotFoundException
+        if (taskService.findTaskById(id)==null) {
+            throw new ResourceNotFoundException(id);
+        }
+        else {
         taskService.deleteTask(id);
         // TaskService states deleteTask(Task task).
-        return ResponseEntity.ok(true);
-        // return ResponseEntity.noContent().build(); // If the task is deleted successfully, returns a ResponseEntity object with a 204 No Content status code.
+        // Because the method returns <Boolean> so keeping the return ResponseEntity.ok(true);
+        return ResponseEntity.ok(true);}
     }
 }
